+++ support-frontend/assets/components/forms/customFields/error.scss
@@ -1,4 +1,4 @@
-@import './formFields';
+@import '../formFields';
 
 .component-form-error__error {
   @include gu-fontset-explainer;
