+++ support-frontend/assets/components/forms/customFields/radioInput.scss
@@ -1,4 +1,4 @@
-@import '../standardFields/formFields';
+@import '../formFields';
 
 .component-radio-input {
   @include gu-fontset-body-sans;
