+++ support-frontend/assets/components/forms/label.scss
