+++ support-frontend/assets/components/forms/legend.scss
