+++ support-frontend/assets/hocs/canShow.jsx
