+++ support-frontend/assets/hocs/withError.jsx
@@ -4,7 +4,7 @@
 
 import React from 'react';
 
-import { Label, type PropsForHoc } from '../standardFields/label';
+import { Error, type PropsForHoc } from 'components/forms/customFields/error';
 
 // ----- Types ----- //
 
@@ -13,18 +13,20 @@ type AugmentedProps<Props> = Props & PropsForHoc;
 type In<Props> = React$ComponentType<Props>;
 type Out<Props> = React$ComponentType<AugmentedProps<Props>>;
 
+
 // ----- Component ----- //
 
-function withLabel<Props: { id: string }>(Component: In<Props>): Out<Props> {
-  return ({
-    label, optional, footer, ...props
-  }: AugmentedProps<Props>) => (
-    <Label htmlFor={props.id} footer={footer} label={label} optional={optional}>
+function withError<Props: { id: string }>(Component: In<Props>): Out<Props> {
+
+  return ({ error, ...props }: AugmentedProps<Props>) => (
+    <Error htmlFor={props.id} error={error}>
       <Component {...props} />
-    </Label>
+    </Error>
   );
+
 }
+
 
 // ----- Exports ----- //
 
-export { withLabel };
+export { withError };
