+++ support-frontend/assets/hocs/withLabel.jsx
@@ -4,7 +4,7 @@
 
 import React from 'react';
 
-import { Error, type PropsForHoc } from '../standardFields/error';
+import { Label, type PropsForHoc } from 'components/forms/label';
 
 // ----- Types ----- //
 
@@ -13,20 +13,18 @@ type AugmentedProps<Props> = Props & PropsForHoc;
 type In<Props> = React$ComponentType<Props>;
 type Out<Props> = React$ComponentType<AugmentedProps<Props>>;
 
-
 // ----- Component ----- //
 
-function withError<Props: { id: string }>(Component: In<Props>): Out<Props> {
-
-  return ({ error, ...props }: AugmentedProps<Props>) => (
-    <Error htmlFor={props.id} error={error}>
+function withLabel<Props: { id: string }>(Component: In<Props>): Out<Props> {
+  return ({
+    label, optional, footer, ...props
+  }: AugmentedProps<Props>) => (
+    <Label htmlFor={props.id} footer={footer} label={label} optional={optional}>
       <Component {...props} />
-    </Error>
+    </Label>
   );
-
 }
-
 
 // ----- Exports ----- //
 
-export { withError };
+export { withLabel };
