+++ support-frontend/assets/pages/paper-subscription-checkout/components/checkoutForm.jsx
@@ -14,15 +14,15 @@ import Rows from 'components/base/rows';
 import Text from 'components/text/text';
 import CheckoutExpander from 'components/checkoutExpander/checkoutExpander';
 import Button from 'components/button/button';
-import { Input } from 'components/forms/standardFields/input';
-import { Select } from 'components/forms/standardFields/select';
-import { Fieldset } from 'components/forms/standardFields/fieldset';
+import { Input } from 'components/forms/input';
+import { Select } from 'components/forms/select';
+import { Fieldset } from 'components/forms/fieldset';
 import { sortedOptions } from 'components/forms/customFields/sortedOptions';
 import { options } from 'components/forms/customFields/options';
 import { RadioInput } from 'components/forms/customFields/radioInput';
-import { withLabel } from 'components/forms/formHOCs/withLabel';
-import { withError } from 'components/forms/formHOCs/withError';
-import { asControlled } from 'components/forms/formHOCs/asControlled';
+import { withLabel } from 'hocs/withLabel';
+import { withError } from 'hocs/withError';
+import { asControlled } from 'hocs/asControlled';
 import Form, { FormSection } from 'components/checkoutForm/checkoutForm';
 import Checkout from 'components/checkout/checkout';
 import GeneralErrorMessage from 'components/generalErrorMessage/generalErrorMessage';
